requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
distlib==0.3.3
Django==3.2.6
django-nested-admin==3.3.3
filelock==3.3.0
gitdb==4.0.7
GitPython==3.1.24
Pillow==8.3.1
platformdirs==2.4.0
python-dotenv==0.19.0
python-monkey-business==1.0.0
pytz==2021.1
six==1.16.0
smmap==4.0.0
sqlparse==0.4.1
typing-extensions==3.10.0.2
virtualenv==20.8.1
